requirements.txt
# --- Core Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# --- Database & Environment ---
mysql-connector-python>=8.3.0
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic-settings>=2.2.1

# --- Utilities ---
Faker>=24.4.0